+++ src/components/CreatePost.js
@@ -1,7 +1,9 @@
 import React, { useState } from "react";
-import Selectbox from "./Selectbox";
 import gql from "graphql-tag";
 import { Mutation } from "react-apollo";
+import { useHistory } from "react-router-dom";
+
+import AuthorSelect from "./AuthorSelect";
 
 const POST_MUTATION = gql`
   mutation addPost($post: PostInput!) {
@@ -19,15 +21,13 @@ const POST_MUTATION = gql`
   }
 `;
 
-export default function Postbox(props) {
+export default function AddPost(props) {
   const [title, setTitle] = useState("");
   const [text, setText] = useState("");
   const [authorId, setAuthorId] = useState("");
 
-  const handleAuthorSelectFromChild = (authorName, authorId) =>
+  const handleChangeAuthor = (authorName, authorId) =>
     setAuthorId(authorId);
-
-  const onChangeText = e => setText(e.target.value);
 
   const onChangeTitle = e => setTitle(e.target.value);
 
@@ -43,11 +43,11 @@ export default function Postbox(props) {
   return (
     <div>
       <div className="flex flex-column mt3">
-        <Selectbox onChange={handleAuthorSelectFromChild} />
+        <AuthorSelect onChange={handleChangeAuthor} />
         <input
           className="mb2"
           value={title}
-          onChange={onChangeTitle}
+          onChange={e => setTitle(e.target.value)}
           type="text"
           placeholder="Add Title"
         />
@@ -56,7 +56,7 @@ export default function Postbox(props) {
           rows="15"
           cols="100"
           value={text}
-          onChange={onChangeText}
+          onChange={e => setText(e.target.value)}
           type="text"
           placeholder="Add your blog post"
         />
